+++ src/components/date-time-picker/date-time-picker-strings.ts
@@ -1,0 +1,12 @@
+export interface DateTimePickerStrings {
+    hours: string;
+    minutes: string;
+    seconds: string;
+    okay: string;
+    cancel: string;
+    months: string[];
+    monthsShort?: string[];
+    weekdays: string[];
+    weekdaysShort?: string[];
+    weekdaysMin?: string[];
+}